# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Propolis Fulfillment API",
        "description": "Dokumentasi resmi layanan backend Propolis. Semua response menggunakan format {success, data, message}.",
        "contact": {
            "name": "Tim Propolis",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Root"
        },
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrasi akun baru",
                "operationId": "ed42f392010c301892e640067141676e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_lengkap",
                                    "username",
                                    "email",
                                    "password",
                                    "province_id",
                                    "province_name",
                                    "city_id",
                                    "city_name",
                                    "district_id",
                                    "district_name",
                                    "subdistrict_id",
                                    "subdistrict_name",
                                    "postal_code",
                                    "alamat_lengkap"
                                ],
                                "properties": {
                                    "nama_lengkap": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "jane"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8
                                    },
                                    "no_hp": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "pelanggan"
                                        ],
                                        "example": "pelanggan"
                                    },
                                    "province_id": {
                                        "type": "integer"
                                    },
                                    "province_name": {
                                        "type": "string"
                                    },
                                    "city_id": {
                                        "type": "integer"
                                    },
                                    "city_name": {
                                        "type": "string"
                                    },
                                    "district_id": {
                                        "type": "integer"
                                    },
                                    "district_name": {
                                        "type": "string"
                                    },
                                    "subdistrict_id": {
                                        "type": "integer"
                                    },
                                    "subdistrict_name": {
                                        "type": "string"
                                    },
                                    "postal_code": {
                                        "type": "string",
                                        "example": "40123"
                                    },
                                    "alamat_lengkap": {
                                        "type": "string",
                                        "example": "Jl. Mawar no 1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registrasi berhasil"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login dan mendapatkan token",
                "operationId": "8cb7742ae98d0990acb2907c4e00bf24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil"
                    },
                    "401": {
                        "description": "Kredensial salah"
                    }
                }
            }
        },
        "/auth/profile": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Mendapatkan informasi profil user",
                "operationId": "19ebac4c2899c32cfe90402c3781e2e6",
                "responses": {
                    "200": {
                        "description": "Data profil user"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Auth"
                ],
                "summary": "Update profil user",
                "operationId": "b94c6da5a01ce99b401b644226460853",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "nama_lengkap": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "no_hp": {
                                        "type": "string"
                                    },
                                    "province_id": {
                                        "type": "integer"
                                    },
                                    "city_id": {
                                        "type": "integer"
                                    },
                                    "district_id": {
                                        "type": "integer"
                                    },
                                    "subdistrict_id": {
                                        "type": "integer"
                                    },
                                    "postal_code": {
                                        "type": "string"
                                    },
                                    "alamat_lengkap": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil berhasil diupdate"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout dan revoke token",
                "operationId": "0d29c740dc06d77602af8095bf4654cc",
                "responses": {
                    "200": {
                        "description": "Logout berhasil"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Token",
                "description": "Masukkan token Sanctum dengan format \"Bearer {token}\"."
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}